monotonicityTest: A test for a monotonic trend as conducted originally in StatCharrms

This is a function originally encoded by Joe Swintek for the R package StatCharrms. Performs the test for monotonicity as described in the OECD guidance No. 54 for statistics in ecotoxicology. This is the test for monotonicity as done in the SAS Version it sets up linear and Quadratic contrast for an ANOVA Uses .stdEndEnv$WeightVar

Usage

monotonicityTest(Data, Treatment, Response)

Arguments

Data

A data set.

Treatment

The name (as a string) of the treatment variable.

Response

The name (as a string) of the response variable.

Value

A table of test statistics for both the linear and quadratic trends.

Details

Calls getLineContrast and getQuadContrast to attain the contrasts and rankTransform for the monotonicity test.

References

https://www.rdocumentation.org/packages/StatCharrms/versions/0.90.96

OECD (2006) Current approaches in the statistical analysis of ecotoxicity data: A guidance to application. Environmental Health and Safety Publications. Series on Testing and Assessment. No. 54. Paris.

Author

Joe Swintek

Examples

data(lengthWeightData)
SubData <- lengthWeightData[lengthWeightData$Age == "8 week", ]
SubData <- SubData[SubData$Generation == "F1", ]
SubData <- SubData[SubData$SEX == "M", ]