rankTransform

This is a function originally encoded by Joe Swintek for the R package StatCharrms. It Transforms a response by rank into critical values of the standard normal distribution. In the case of ties this function will use the mean of the transformed response. This function will rank transform the data the same way the SAS code below does: proc rank data=sorted out=assumtst normal=blom ties=mean; var Counts; by generation; ranks rank_EGGS; run;

Usage

rankTransform(Data, VecName)

Arguments

Data

A dataset

VecName

The name (as a string) of the response to be transformed.

Value

The original data set with a new variable called 'TransformedResponse' which is the rank transform of the response.

References

https://www.rdocumentation.org/packages/StatCharrms/versions/0.90.96

Author

Joe Swintek

Examples

data(lengthWeightData)
SubData <- lengthWeightData[lengthWeightData$Age == "16 week", ]
SubData <- SubData[SubData$Generation == "F1", ]
SubData <- SubData[SubData$SEX == "M", ]
RankData <- rankTransform(Data = SubData, VecName = "WEIGHT")
head(RankData)
#>     Generation Treatment Replicate Number SEX WEIGHT LENGTH Concentration
#> 317         F1         6         E      2   M    142     23           102
#> 42          F1         2         B      4   M    205     30             6
#> 28          F1         1         C     10   M    208     30             0
#> 182         F1         1         F      5   M    248     31             0
#> 30          F1         2         A      2   M    268     31             6
#> 271         F1         4         F      8   M    277     33            25
#>         Age TransformedResponse
#> 317 16 week          -2.0913539
#> 42  16 week          -1.6701458
#> 28  16 week          -1.4280250
#> 182 16 week          -1.2489624
#> 30  16 week          -1.1028945
#> 271 16 week          -0.9212647